5 acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,14 @@
 class QuestionsController < ApplicationController
-
+  before_action :authenticate_user!, except: %i[index show]
   #before_action :load_question, only: %i[show edit update destroy]
 
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @question = Question.find(params[:id])
+  end
 
   def new
     @question = Question.new
@@ -15,10 +17,10 @@ class QuestionsController < ApplicationController
   def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -34,7 +36,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    question.destroy if current_user.author?(question)
     redirect_to questions_path
   end
